requirements.txt
alpaca-py==0.43.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
beautifulsoup4==4.14.2
caffeine==0.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
fastapi==0.121.0
fonttools==4.60.1
frozendict==2.4.6
h11==0.16.0
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
msgpack==1.1.2
multitasking==0.0.12
numpy==2.3.4
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sseclient-py==1.8.0
starlette==0.49.3
supabase==2.24.0
ta==0.11.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yfinance==0.2.66
